Mac: Always write shell scripts as UTF-8
Always write shell scripts in UTF-8 encoding because nl_langinfo doesn't
return anything useful when the application is started via Finder. Most
likely LANG and LC_CTYPE environment variables are only set in Terminal
shells but not in the rest of the operating system.

This oversight caused files to be written in UTF-8 still but the header
and ZEIT_ENCODING variable would state that it's US-ASCII (a heinous
lie).

Alternatively we could muck about with QTextCodec::codecForLocale() to
write the files in some different, system-wide encoding. But Mac systems
rarely run with anything other than UTF-8 (including Terminal shells).
Also, Windows already uses UTF-8 statically and the Mac has supported
and used UTF-8 for everything for ages. So I don't see the benefit.