mac: Fix right mouse button for good (hopefully)
The mac port exposed early on that using the context menu logic to
intercept right button presses is horribly broken. We used to patch Qt
to work around this but that never fully solved the problem.

Additional debugging now shows that signal customContextMenuRequested()
fires on the mouse button press already and does not wait for the
release. This causes the event handling and mouse button state tracking
inside Qt to become corrupted. This in turn causes "ghost" button
presses to happen once the account dialog is closed again and actual
button presses to be ignored.

It is unclear why this only happens on the Mac. From looking at the Qt
code this should also affect other OSes.

This change switches to actual tracking of right mouse button presses
and matching releases and emitting a custom itemRightClicked() signal.

Looking at the Qt code also assures us that itemClicked() only fires on
left button clicks and makes sure that the item delivered to the slot is
valid. This allows us to get rid of some other cruft such as
getCurrentButton().