mac: Update dependencies and fix SDKs
Update dependencies to recent versions. Drop patches which became
unnecessary.

PostgreSQL now requires PG_SYSROOT to be set to correctly inject a
specific SDK to use into its build system. Also it uses MAKELEVEL to get
its internal dependencies on generated headers right which we need to
work around in our Makefile.

While at it, sync the SDK version used for dependencies and scTime
itself and set it to 10.12.